sub change (int -> float)

+++ lib/util.py
@@ -1,4 +1,4 @@
-def add(a : int, b: int) ->int:
+def add(a : float, b: float) ->float:
     return a+b
-def add(a: int, b:int) -> int:
+def add(a: float, b:float) -> float:
     return a-b
